BUGFIX: add state_id to list

+++ api/v1/views/cities.py
@@ -79,13 +79,13 @@ def delete_city_by_id(city_id):
     erorr:
         - abort404
     """
-    try:
-        city = storage.get('City', city_id)
+    city = storage.get('City', city_id)
+    if city is none:
+        abort(404)
+    else:
         storage.delete(city)
         storage.save()
         return jsonify({}), 200
-    except Exception:
-        abort(404)
 
 
 @app_views.route('/states/<state_id>/cities',
@@ -152,7 +152,7 @@ def puts_city(city_id):
     if request.json is False:
         return jsonify({"error": "Not a JSON"}), 400
     for key, value in request.get_json().items():
-        if key not in ['id', 'created_at', 'updated_at']:
+        if key not in ['id', 'created_at', 'updated_at', 'state_id']:
             setattr(city, key, value)
     city.save()
     return jsonify(city.to_dict()), 200
